+++ apps/ui/src/data/queries/use-sync-site.ts
@@ -1,38 +1,14 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useConnector } from '@/data/core';
 import { connectedWpcomSitesQueryKey } from '@/data/queries/use-connected-wpcom-sites';
-import { SNAPSHOTS_QUERY_KEY } from '@/data/queries/use-snapshots';
+import { SITES_QUERY_KEY } from '@/data/queries/use-sites';
 import { reportSyncError, reportSyncPending, reportSyncSuccess } from '@/data/sync-activity';
 
-// Exported so other mutation hooks (e.g. a future pull) and the sync-activity
-// indicator can line up on the same keys when filtering the mutation cache.
+// Mutation keys are exported so downstream consumers (e.g. a cross-page
+// activity indicator or future bulk-sync UI) can filter the react-query
+// mutation cache for in-flight push/pull operations by site.
 export const PUSH_TO_LIVE_MUTATION_KEY = [ 'pushSiteToLive' ] as const;
 export const PULL_FROM_LIVE_MUTATION_KEY = [ 'pullSiteFromLive' ] as const;
-
-type PublishPreviewVariables = {
-	siteId: string;
-	existingHostname?: string;
-};
-
-export function usePublishPreviewSite() {
-	const connector = useConnector();
-	const queryClient = useQueryClient();
-	return useMutation( {
-		mutationFn: ( { siteId, existingHostname }: PublishPreviewVariables ) =>
-			connector.publishPreviewSite( siteId, existingHostname ),
-		onMutate: ( { siteId } ) => {
-			reportSyncPending( siteId, 'preview' );
-		},
-		onSuccess: ( _result, { siteId } ) => {
-			reportSyncSuccess( siteId, 'preview' );
-			void queryClient.invalidateQueries( { queryKey: SNAPSHOTS_QUERY_KEY } );
-		},
-		onError: ( error, { siteId } ) => {
-			const message = error instanceof Error ? error.message : String( error );
-			reportSyncError( siteId, 'preview', message );
-		},
-	} );
-}
 
 type PushToLiveVariables = {
 	siteId: string;
@@ -82,7 +58,7 @@ export function usePullSiteFromLive() {
 			// The CLI may have stopped/started the server during the import,
 			// and the site's database + themes just changed — refresh the
 			// site list so any downstream consumers see the new state.
-			void queryClient.invalidateQueries( { queryKey: [ 'sites' ] } );
+			void queryClient.invalidateQueries( { queryKey: SITES_QUERY_KEY } );
 		},
 		onError: ( error, { siteId } ) => {
 			const message = error instanceof Error ? error.message : String( error );
